Changed functionality of main game page
Changed "main.html" to "main.php", added "Previous" and "Next" links for navigation through the story. Also added a "contid" variable for the users table, so we can tell how far the user is in the story. Comments can be found in "main.php". Also made minor changes to some other files